requirements.txt
Django==5.0.6
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.15.1
pillow==10.3.0
python-dotenv==1.0.1
boto3==1.35.9
django-widgets-improved==1.5.0
mysqlclient==2.1.0
django-allauth==65.1.0
django-channels==0.7.0
cloudinary==1.41.0